distribute/requirements.txt
iconsdk==1.3.2
tbears==1.7.1
python-telegram-bot==12.8
